feat: replace the simple central state with local state.
This would have been fairly simple if it wasn't for the range slider.

The slider interaction isn't very smooth if it's rerendering the element
while you are dragging it. This meant we needed a way to make the slider
uncontrolled and the other form elements controlled.

This is doable, but the LengthSlider was squarely placed within the
controls box so we couldn't easily overwrite innerHTML for the other
components and not the slider.

I ended up distinguishing between initial instantiate and rerenders, so
I could just rerender the relevant components on App.render() (instead
of reassigning innerHTML). This was good, but it meant I needed a way to
pass props on rerenders.

I got it all working but it looks kind of clunky and I'm left thinking
that there's probably a better way to handle it.